Move the HResult-like-struct-as-HRESULT behavior to be COM only
